+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add ComboBoxBehavior to define of max text length of the combobox if it is editable.
+### Supported .Net Versions
+- .Net 6
+- .Net 7
+- .Net 8
 
 ## [1.0.0] - 2023-12-24
 ### Added